+++ routers/base.go
@@ -15,8 +15,10 @@
 package routers
 
 import (
+	"encoding/json"
 	"fmt"
 	"net"
+	"net/http"
 	"net/url"
 	"strings"
 
@@ -32,6 +34,14 @@ type Response struct {
 	Msg    string      `json:"msg"`
 	Data   interface{} `json:"data"`
 	Data2  interface{} `json:"data2"`
+}
+
+type Object struct {
+	Owner        string `json:"owner"`
+	Name         string `json:"name"`
+	Organization string `json:"organization"`
+	AccessKey    string `json:"accessKey"`
+	AccessSecret string `json:"accessSecret"`
 }
 
 func responseError(ctx *context.Context, error string, code int, data ...interface{}) {
@@ -84,6 +94,30 @@ func getUsernameByClientIdSecret(ctx *context.Context) string {
 	}
 
 	return fmt.Sprintf("app/%s", application.Name)
+}
+
+func getKeys(ctx *context.Context) (string, string) {
+	method := ctx.Request.Method
+
+	if method == http.MethodGet {
+		accessKey := ctx.Input.Query("accessKey")
+		accessSecret := ctx.Input.Query("accessSecret")
+		return accessKey, accessSecret
+	} else {
+		body := ctx.Input.RequestBody
+
+		if len(body) == 0 {
+			return ctx.Request.Form.Get("accessKey"), ctx.Request.Form.Get("accessSecret")
+		}
+
+		var obj Object
+		err := json.Unmarshal(body, &obj)
+		if err != nil {
+			return "", ""
+		}
+
+		return obj.AccessKey, obj.AccessSecret
+	}
 }
 
 func getUsernameByKeys(ctx *context.Context) string {
